refactor: InterviewSetupModal 컴포넌트 코드 및 컴포넌트 분리
